working in checkout login out
